v0.0.2.29

+++ Mythic_CLI/src/cmd/internal/installservice.go
@@ -149,28 +149,39 @@ func InstallFolder(installPath string, overWrite bool) error {
 				for _, f := range files {
 					if f.IsDir() {
 						fmt.Printf("[*] Processing Documentation for %s\n", f.Name())
-						if dirExists(filepath.Join(workingPath, "documentation-docker", "content", "Agents", f.Name())) {
-							if overWrite || askConfirm("[*] "+f.Name()+" documentation already exists. Replace current version? ") {
-								fmt.Printf("[*] Removing current version\n")
-								err = os.RemoveAll(filepath.Join(workingPath, "documentation-docker", "content", "Agents", f.Name()))
-								if err != nil {
-									fmt.Printf("[-] Failed to remove current version: %v\n", err)
-									fmt.Printf("[-] Continuing to the next payload documentation\n")
-									continue
+						if mythicEnv.GetBool("documentation_bind_local_mount") {
+							if dirExists(filepath.Join(workingPath, "documentation-docker", "content", "Agents", f.Name())) {
+								if overWrite || askConfirm("[*] "+f.Name()+" documentation already exists. Replace current version? ") {
+									fmt.Printf("[*] Removing current version\n")
+									err = os.RemoveAll(filepath.Join(workingPath, "documentation-docker", "content", "Agents", f.Name()))
+									if err != nil {
+										fmt.Printf("[-] Failed to remove current version: %v\n", err)
+										fmt.Printf("[-] Continuing to the next payload documentation\n")
+										continue
+									} else {
+										fmt.Printf("[+] Successfully removed the current version\n")
+									}
 								} else {
-									fmt.Printf("[+] Successfully removed the current version\n")
+									fmt.Printf("[!] Skipping documentation for , %s\n", f.Name())
+									continue
 								}
-							} else {
-								fmt.Printf("[!] Skipping documentation for , %s\n", f.Name())
+							}
+							fmt.Printf("[*] Copying new documentation into place\n")
+							err = copyDir(filepath.Join(installPath, "documentation-payload", f.Name()), filepath.Join(workingPath, "documentation-docker", "content", "Agents", f.Name()))
+							if err != nil {
+								fmt.Printf("[-] Failed to copy directory over\n")
+								continue
+							}
+						} else {
+							err = moveFileToVolume("mythic_documentation_volume",
+								filepath.Join("content", "Agents", f.Name()),
+								filepath.Join(installPath, "documentation-payload", f.Name()))
+							if err != nil {
+								fmt.Printf("[-] Failed to install documentation for payload: %v\n", err)
 								continue
 							}
 						}
-						fmt.Printf("[*] Copying new documentation into place\n")
-						err = copyDir(filepath.Join(installPath, "documentation-payload", f.Name()), filepath.Join(workingPath, "documentation-docker", "content", "Agents", f.Name()))
-						if err != nil {
-							fmt.Printf("[-] Failed to copy directory over\n")
-							continue
-						}
+
 					}
 				}
 				fmt.Printf("[+] Successfully installed Payload documentation\n")
